requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
numpy==1.18.5
Pillow==8.0.0
PyJWT==1.7.1
sqlparse==0.4.1
psycopg2-binary==2.8.6
vncorenlp==1.0.3
pyOpenSSL==19.1.0
django-sslserver==0.22
timeloop==1.0.2
pickle-mixin==1.0.2
pandas==1.1.3
channels==3.0.0
asgiref==3.2.10
channels-redis==3.2.0
daphne==3.0.0
tensorflow==2.3.0
transformers==3.2.0
scikit-learn==0.23.2
pandas==1.0.5
psutil==5.7.3